Show the update banner everywhere
